refactor(ui): split components and consolidate utilities
- Extract shared tool-call utilities (formatArguments, getResultPreview, etc.)
- Split CapabilitySelector (619 lines) into focused components:
  - CapabilityDialog: selection dialog with search/categories
  - SelectedCapabilityList: ordered list of selected capabilities
  - useCapabilityDependencies: hook for dependency logic
- Update capabilities spec with:
  - Quick reference mapping table for built-in capabilities
  - Clarified capability ID format (built-in vs MCP)
  - Enhanced MCP section with tool naming details
